Feat: Payment Controller Feature added

+++ src/main/java/com/smart/event/event_management_system/service/PaymentService.java
@@ -3,5 +3,5 @@ package com.smart.event.event_management_system.service;
 import com.smart.event.event_management_system.entity.Payment;
 
 public interface PaymentService {
-    Payment makePayment(Integer bookingId, String mode);
+    Payment makePayment(Long bookingId, String mode);
 }
